+++ src/main/kotlin/io/ionic/libs/iongeolocationlib/controller/helper/IONGLOCGoogleServicesHelper.kt
@@ -1,13 +1,17 @@
-package io.ionic.libs.iongeolocationlib.controller
+package io.ionic.libs.iongeolocationlib.controller.helper
 
 import android.annotation.SuppressLint
 import android.app.Activity
 import android.location.Location
+import android.location.LocationManager
+import android.net.ConnectivityManager
 import android.os.Looper
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.IntentSenderRequest
+import androidx.core.location.LocationManagerCompat
 import com.google.android.gms.common.ConnectionResult
 import com.google.android.gms.common.GoogleApiAvailability
+import com.google.android.gms.common.api.ApiException
 import com.google.android.gms.common.api.ResolvableApiException
 import com.google.android.gms.location.CurrentLocationRequest
 import com.google.android.gms.location.FusedLocationProviderClient
@@ -18,12 +22,15 @@ import com.google.android.gms.location.LocationSettingsRequest
 import com.google.android.gms.location.Priority
 import io.ionic.libs.iongeolocationlib.model.IONGLOCException
 import io.ionic.libs.iongeolocationlib.model.IONGLOCLocationOptions
+import io.ionic.libs.iongeolocationlib.model.internal.LocationSettingsResult
 import kotlinx.coroutines.tasks.await
 
 /**
- * Helper class that wraps the functionality of FusedLocationProviderClient
+ * Helper class that wraps the functionality of [FusedLocationProviderClient]
  */
-class IONGLOCServiceHelper(
+internal class IONGLOCGoogleServicesHelper(
+    private val locationManager: LocationManager,
+    private val connectivityManager: ConnectivityManager,
     private val fusedLocationClient: FusedLocationProviderClient,
     private val activityLauncher: ActivityResultLauncher<IntentSenderRequest>
 ) {
@@ -31,20 +38,20 @@ class IONGLOCServiceHelper(
      * Checks if location is on, as well as other conditions for retrieving device location
      * @param activity the Android activity from which the location request is being triggered
      * @param options location request options to use
-     * @param interval interval for requesting location updates; use 0 if meant to retrieve a single location
-     * @return true if location was checked and is on, false if it requires user to resolve issue (e.g. turn on location)
-     *          If false, the result is returned in `resolveLocationSettingsResultFlow`
+     * @param shouldTryResolve true if should try to resolve errors; false otherwise.
+     * Dictates whether [LocationSettingsResult.Resolving] or [LocationSettingsResult.ResolveSkipped] is returned.
+     * The exception being if location is off, in which case it will always be resolved.
+     * @return result of type [LocationSettingsResult]
      * @throws [IONGLOCException.IONGLOCSettingsException] if an error occurs that is not resolvable by user
      */
     internal suspend fun checkLocationSettings(
         activity: Activity,
         options: IONGLOCLocationOptions,
-        interval: Long
-    ): Boolean {
-
+        shouldTryResolve: Boolean
+    ): LocationSettingsResult {
         val request = LocationRequest.Builder(
             if (options.enableHighAccuracy) Priority.PRIORITY_HIGH_ACCURACY else Priority.PRIORITY_BALANCED_POWER_ACCURACY,
-            interval
+            options.interval
         ).build()
 
         val builder = LocationSettingsRequest.Builder()
@@ -53,37 +60,43 @@ class IONGLOCServiceHelper(
 
         try {
             client.checkLocationSettings(builder.build()).await()
-            return true
+            return LocationSettingsResult.Success
         } catch (e: ResolvableApiException) {
-
-            // Show the dialog to enable location by calling startResolutionForResult(),
-            // and then handle the result in onActivityResult
-            val resolutionBuilder: IntentSenderRequest.Builder =
-                IntentSenderRequest.Builder(e.resolution)
-            val resolution: IntentSenderRequest = resolutionBuilder.build()
-
-            activityLauncher.launch(resolution)
+            val locationOn = LocationManagerCompat.isLocationEnabled(locationManager)
+            if (!locationOn || shouldTryResolve) {
+                // Show the dialog to enable location by calling startResolutionForResult(),
+                // and then handle the result in onActivityResult
+                val resolutionBuilder: IntentSenderRequest.Builder =
+                    IntentSenderRequest.Builder(e.resolution)
+                val resolution: IntentSenderRequest = resolutionBuilder.build()
+                activityLauncher.launch(resolution)
+                return LocationSettingsResult.Resolving
+            } else {
+                return LocationSettingsResult.ResolveSkipped(e)
+            }
         } catch (e: Exception) {
-            throw IONGLOCException.IONGLOCSettingsException(
-                message = "There is an error with the location settings.",
-                cause = e
-            )
+            return LocationSettingsResult.UnresolvableError(e.mapLocationSettingsError())
         }
-        return false
     }
 
     /**
      * Checks if the device has google play services, required to use [FusedLocationProviderClient]
      * @param activity the Android activity from which the location request is being triggered
-     * @return true if google play services is available, false otherwise
+     * @param shouldTryResolve true if should try to resolve errors; false otherwise.
+     * @return Success if google play services is available, Error otherwise
      */
-    internal fun checkGooglePlayServicesAvailable(activity: Activity): Result<Unit> {
+    internal fun checkGooglePlayServicesAvailable(
+        activity: Activity,
+        shouldTryResolve: Boolean
+    ): Result<Unit> {
         val googleApiAvailability = GoogleApiAvailability.getInstance()
         val status = googleApiAvailability.isGooglePlayServicesAvailable(activity)
 
         return if (status != ConnectionResult.SUCCESS) {
             if (googleApiAvailability.isUserResolvableError(status)) {
-                googleApiAvailability.getErrorDialog(activity, status, 1)?.show()
+                if (shouldTryResolve) {
+                    googleApiAvailability.getErrorDialog(activity, status, 1)?.show()
+                }
                 sendResultWithGoogleServicesException(
                     resolvable = true,
                     message = "Google Play Services error user resolvable."
@@ -97,23 +110,6 @@ class IONGLOCServiceHelper(
         } else {
             Result.success(Unit)
         }
-    }
-
-    /**
-     * Returns a Result object containing an IONGLOCException.IONGLOCGoogleServicesException exception with the given
-     * resolvable and message values
-     * @param resolvable whether or not the exception is resolvable
-     * @param message message to include in the exception
-     * @return Result object with the exception to return
-     *
-     */
-    private fun sendResultWithGoogleServicesException(resolvable: Boolean, message: String): Result<Unit> {
-        return Result.failure(
-            IONGLOCException.IONGLOCGoogleServicesException(
-                resolvable = resolvable,
-                message = message
-            )
-        )
     }
 
     /**
@@ -148,7 +144,7 @@ class IONGLOCServiceHelper(
         options: IONGLOCLocationOptions,
         locationCallback: LocationCallback
     ) {
-        val locationRequest = LocationRequest.Builder(options.timeout).apply {
+        val locationRequest = LocationRequest.Builder(options.interval).apply {
             setMaxUpdateAgeMillis(options.maximumAge)
             setPriority(if (options.enableHighAccuracy) Priority.PRIORITY_HIGH_ACCURACY else Priority.PRIORITY_BALANCED_POWER_ACCURACY)
             if (options.minUpdateInterval != null) {
@@ -174,5 +170,25 @@ class IONGLOCServiceHelper(
         locationCallback: LocationCallback
     ) {
         fusedLocationClient.removeLocationUpdates(locationCallback)
+    }
+
+    /**
+     * Map the Location Settings Exception to an exception from this native library.
+     * @return a [IONGLOCException]
+     */
+    private fun Exception.mapLocationSettingsError(): IONGLOCException = if (this is ApiException &&
+        message?.contains("SETTINGS_CHANGE_UNAVAILABLE") == true
+        && !LocationManagerCompat.isLocationEnabled(locationManager)
+        && !hasNetworkEnabledForLocationPurposes(locationManager, connectivityManager)
+    ) {
+        IONGLOCException.IONGLOCLocationAndNetworkDisabledException(
+            message = "Unable to retrieve location because device has both Network and Location turned off.",
+            cause = this
+        )
+    } else {
+        IONGLOCException.IONGLOCSettingsException(
+            message = "There is an error with the location settings.",
+            cause = this
+        )
     }
 }