+++ CHANGELOG.md
@@ -4,6 +4,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0]
+
+### 2026-02-02
+- Feature: Add support for magnetic and true heading using device sensors (available for location updates only).
+- Feature: Add `magneticHeading`, `trueHeading`, `headingAccuracy`, and `course` to `IONGLOCLocationResult`.
+
+## [2.1.0]
+
+### 2025-10-31
+
+- Feature: Add native timeout to `watchPosition` Add new `interval` to control location updates without `timeout` variable.
+
+## [2.0.0]
+
+### 2025-09-30
+
+- Feature: Allow using a fallback if Google Play Services fails.
+
+BREAKING CHANGE: The constructor for the controller and some of its methods have changed signature.
+You will need to change how your application calls the library if you update to this version.
+
+### 2025-06-26
+
+- Migrate publishing from OSSRH to Central Portal.
+
 ## [1.0.0]
 
 ### 2025-01-20